test(code-sandbox): avoid duplicate app logger setup
